+++ pkg/kubelet/dockershim/network/hostport/hostport.go
@@ -28,7 +28,7 @@ import (
 	"k8s.io/klog/v2"
 
 	v1 "k8s.io/api/core/v1"
-	utiliptables "k8s.io/kubernetes/pkg/util/iptables"
+	utiliptables "k8s.io/component-helpers/network/utils/iptables"
 )
 
 const (
