+++ staging/src/k8s.io/component-helpers/go.mod
@@ -6,6 +6,7 @@ go 1.16
 
 require (
 	github.com/google/go-cmp v0.5.5
+	golang.org/x/sys v0.0.0-20210820121016-41cdb8703e55
 	k8s.io/api v0.0.0
 	k8s.io/apimachinery v0.0.0
 	k8s.io/client-go v0.0.0
