+++ staging/src/k8s.io/component-helpers/network/utils/iptables/doc.go
@@ -15,4 +15,4 @@ limitations under the License.
 */
 
 // Package iptables provides an interface and implementations for running iptables commands.
-package iptables // import "k8s.io/kubernetes/pkg/util/iptables"
+package iptables // import "k8s.io/component-helpers/network/utils/iptables"
